backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
werkzeug==3.0.3
pycryptodome==3.20.0
gunicorn==21.2.0
cryptography==43.0.1
requests==2.32.3
scikit-learn==1.5.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
flask-migrate==4.0.4
docker